third_party/blink/web_tests/editing/pasteboard/drop-text-events-expected.txt

This tests that Drag drop fires textInput events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS actualTextEventData is 'PlainHello'
PASS testTargetTextarea.value is 'PlainHello'
PASS actualTextEventData is 'PlainHello'
PASS testTargetInput.value is 'PlainHello'
PASS actualTextEventData is ''
PASS 0 <= testTargetEditable.innerHTML.indexOf('PlainHello') is true
PASS actualTextEventData is 'RichHello'
PASS testTargetTextarea.value is 'RichHello'
PASS actualTextEventData is 'RichHello'
PASS testTargetInput.value is 'RichHello'
PASS actualTextEventData is ''
PASS 0 <= testTargetEditable.innerHTML.indexOf('<b>Rich</b>Hello') is true
PASS textInputCount is proceedingTestCases.length
PASS actualTextEventData is 'PlainHello'
PASS testTargetTextarea.value is ''
PASS actualTextEventData is 'PlainHello'
PASS testTargetInput.value is ''
PASS actualTextEventData is ''
PASS testTargetEditable.innerHTML is 'placeholder'
PASS actualTextEventData is 'RichHello'
PASS testTargetTextarea.value is ''
PASS actualTextEventData is 'RichHello'
PASS testTargetInput.value is ''
PASS actualTextEventData is ''
PASS testTargetEditable.innerHTML is 'placeholder'
PASS textInputCount is cancelingTestCases.length
PASS successfullyParsed is true

TEST COMPLETE
